Update SDL2 and SDL2 Android code
